--- models/image_test.go
@@ -1,7 +1,0 @@
-package models
-
-import "testing"
-
-func Test_Image(t *testing.T) {
-	t.Fatal("This test needs to be implemented!")
-}
