--- models/init.go
@@ -1,7 +1,0 @@
-package models
-
-import (
-	"github.com/technosophos/moniker"
-)
-
-var namer = moniker.New()
