--- models/models_test.go
@@ -1,16 +1,0 @@
-package models_test
-
-import (
-	"testing"
-
-	"github.com/gobuffalo/suite"
-)
-
-type ModelSuite struct {
-	*suite.Model
-}
-
-func Test_ModelSuite(t *testing.T) {
-	as := &ModelSuite{suite.NewModel()}
-	suite.Run(t, as)
-}
