--- models/users/address.go
@@ -1,21 +1,0 @@
-package users
-
-import (
-	"fmt"
-
-	"github.com/jinzhu/gorm"
-)
-
-type Address struct {
-	gorm.Model
-	UserID      uint
-	ContactName string `form:"contact-name"`
-	Phone       string `form:"phone"`
-	City        string `form:"city"`
-	Address1    string `form:"address1"`
-	Address2    string `form:"address2"`
-}
-
-func (address Address) Stringify() string {
-	return fmt.Sprintf("%v, %v, %v", address.Address2, address.Address1, address.City)
-}
