--- models/users/user.go
@@ -1,51 +1,0 @@
-package users
-
-import (
-	"time"
-
-	"github.com/jinzhu/gorm"
-	"github.com/qor/media"
-	"github.com/qor/media/oss"
-)
-
-type User struct {
-	gorm.Model
-	Email    string `form:"email"`
-	Password string
-	Name     string `form:"name"`
-	Gender   string
-	Role     string
-	Birthday *time.Time
-	Avatar   AvatarImageStorage `media_library:"url:/system/{{class}}/{{primary_key}}/{{column}}.{{extension}};path:./public"`
-
-	// Confirm
-	ConfirmToken string
-	Confirmed    bool
-
-	// Recover
-	RecoverToken       string
-	RecoverTokenExpiry *time.Time
-
-	// Accepts
-	AcceptPrivate bool `form:"accept-private"`
-	AcceptLicense bool `form:"accept-license"`
-	AcceptNews    bool `form:"accept-news"`
-}
-
-func (user User) DisplayName() string {
-	return user.Email
-}
-
-func (user User) AvailableLocales() []string {
-	return []string{"en-US", "zh-CN"}
-}
-
-type AvatarImageStorage struct{ oss.OSS }
-
-func (AvatarImageStorage) GetSizes() map[string]*media.Size {
-	return map[string]*media.Size{
-		"small":  {Width: 50, Height: 50},
-		"middle": {Width: 120, Height: 120},
-		"big":    {Width: 320, Height: 320},
-	}
-}
